requirements.txt
aiogram==3.7.0
aiohttp==3.9.5
pydantic==2.7.4
Pygments==2.18.0
python-dotenv==1.0.1
yt_dlp==2024.5.27
